# webui/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateResult": {
        "description": "实际建出来的可能比请求的少：重复下载与需要二次解析的会被拦掉",
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "requested": {
            "title": "Requested",
            "type": "integer"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "requested",
          "created",
          "tasks"
        ],
        "title": "CreateResult",
        "type": "object"
      },
      "CreateTasksRequest": {
        "properties": {
          "episodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 2000,
            "minItems": 1,
            "title": "Episodes",
            "type": "array"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "required": [
          "episodes"
        ],
        "title": "CreateTasksRequest",
        "type": "object"
      },
...
      "DirectoryListing": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "relative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "path",
          "entries",
          "truncated"
        ],
        "title": "DirectoryListing",
        "type": "object"
      },
...
      "FileEntry": {
        "properties": {
          "is_dir": {
            "title": "Is Dir",
            "type": "boolean"
          },
          "is_link": {
            "title": "Is Link",
            "type": "boolean"
          },
          "modified": {
            "title": "Modified",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "is_dir",
          "size",
          "modified",
          "is_link"
        ],
        "title": "FileEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MakeDirResult": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "relative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative"
          }
        },
        "required": [
          "path"
        ],
        "title": "MakeDirResult",
        "type": "object"
      },
      "MkdirRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "name"
        ],
        "title": "MkdirRequest",
        "type": "object"
      },
...
      "TaskSnapshot": {
        "description": "全量快照\n\n`cursor` 是发快照那一刻的事件编号，前端带着它连 WebSocket（`?since=cursor`），\n否则快照与增量之间会漏掉一段",
        "properties": {
          "completed": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Completed",
            "type": "array"
          },
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "downloading": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Downloading",
            "type": "array"
          }
        },
        "required": [
          "cursor",
          "downloading",
          "completed"
        ],
        "title": "TaskSnapshot",
        "type": "object"
      },
      "TaskView": {
        "description": "与 `web/download/view.py` 的 `task_view()` 一一对应\n\n**两边字段必须同步**：那边加了字段这里不加，前端就看不见；\n这里多写一个那边没有的，前端会拿到 undefined",
        "properties": {
          "audio_quality": {
            "default": "",
            "title": "Audio Quality",
            "type": "string"
          },
          "completed_time": {
            "default": 0,
            "title": "Completed Time",
            "type": "integer"
          },
          "cover_id": {
            "default": "",
            "title": "Cover Id",
            "type": "string"
          },
          "created_time": {
            "default": 0,
            "title": "Created Time",
            "type": "integer"
          },
          "download_path": {
            "default": "",
            "title": "Download Path",
            "type": "string"
          },
          "downloaded_size": {
            "default": 0,
            "title": "Downloaded Size",
            "type": "integer"
          },
          "duration": {
            "default": 0,
            "title": "Duration",
            "type": "integer"
          },
          "file_name": {
            "default": "",
            "title": "File Name",
            "type": "string"
          },
          "folder": {
            "default": "",
            "title": "Folder",
            "type": "string"
          },
          "info_label": {
            "default": "",
            "title": "Info Label",
            "type": "string"
          },
          "progress": {
            "default": 0,
            "title": "Progress",
            "type": "integer"
          },
          "speed": {
            "default": 0,
            "title": "Speed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_label": {
            "default": "",
            "title": "Status Label",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_size": {
            "default": 0,
            "title": "Total Size",
            "type": "integer"
          },
          "type": {
            "items": {
              "type": "string"
            },
            "title": "Type",
            "type": "array"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          },
          "video_codec": {
            "default": "",
            "title": "Video Codec",
            "type": "string"
          },
          "video_quality": {
            "default": "",
            "title": "Video Quality",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "status"
        ],
        "title": "TaskView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Bili23 Downloader WebUI",
    "version": "2.15.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/files/list": {
      "get": {
        "description": "列出一个目录。`path` 为空时给第一个根目录",
        "operationId": "list_directory_api_files_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "default": "",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dirs_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dirs Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryListing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Directory",
        "tags": [
          "files"
        ]
      }
    },
    "/api/files/mkdir": {
      "post": {
        "description": "在允许的目录下新建一个子目录\n\n`name` 只能是单个名字，不能是路径：允许它带分隔符的话，`../../x` 就又回来了。\n这里既拦名字本身，也仍然让结果过一遍 `resolve_within_roots`",
        "operationId": "make_directory_api_files_mkdir_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MkdirRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeDirResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Make Directory",
        "tags": [
          "files"
        ]
      }
    },
...
    "/api/tasks": {
      "get": {
        "description": "全量快照\n\n带上 `cursor`：前端拿它去连 WebSocket（`?since=cursor`），服务端会把这之后的\n增量补发出来。没有这个游标的话，快照与增量之间会漏掉一段",
        "operationId": "task_snapshot_api_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Snapshot",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "创建下载任务\n\n`task_manager.create()` 内部会做**重复下载判定**与**是否需要二次解析**的判断，\n所以实际建出来的可能比传进来的少 —— 返回里如实给出两个数字，\n不然前端会以为「点了没反应」",
        "operationId": "create_tasks_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTasksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}